Fix pip install, some restructuring
