+++ bertopic/__init__.py
@@ -1,11 +1,7 @@
 from bertopic._bertopic import BERTopic
-from bertopic._ctfidf import ClassTFIDF
-from bertopic._embeddings import languages
 
 __version__ = "0.6.0"
 
 __all__ = [
     "BERTopic",
-    "ClassTFIDF",
-    "languages"
 ]
