+++ bertopic/backend/_base.py
