+++ bertopic/backend/_sentencetransformers.py
@@ -1,3 +1,21 @@
+from ._base import BaseEmbedder
+from sentence_transformers import SentenceTransformer
+
+
+class SentenceTransformerBackend(BaseEmbedder):
+    def __init__(self, embedding_model):
+        super().__init__()
+
+        if isinstance(embedding_model, SentenceTransformer):
+            self.embedding_model = embedding_model
+        elif isinstance(embedding_model, str):
+            self.embedding_model = SentenceTransformer(embedding_model)
+
+    def embed(self, documents, verbose):
+        embeddings = self.embedding_model.encode(documents, show_progress_bar=verbose)
+        return embeddings
+
+
 languages = ['afrikaans', 'albanian', 'amharic', 'arabic', 'armenian', 'assamese',
              'azerbaijani', 'basque', 'belarusian', 'bengali', 'bengali romanize',
              'bosnian', 'breton', 'bulgarian', 'burmese', 'burmese zawgyi font', 'catalan',
