+++ bertopic/backend/_spacy.py
